# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AfterSale Copilot — Local Simulation",
    "description": "Stages 1–3: single Agent with deterministic after-sales controls. No real payments. Business APIs require authenticated sessions and CSRF. Amounts are CNY strings.",
    "version": "0.3.0"
  },
  "paths": {
...
    "/agent/status": {
      "get": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Status",
        "operationId": "status_agent_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentStatus"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          }
        }
      }
    },
    "/agent/runs": {
      "post": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Chat",
        "operationId": "chat_agent_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            }
          }
        }
      }
    },
    "/agent/runs/{run_id}": {
      "get": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Get Run",
        "operationId": "get_run_agent_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/agent/runs/{run_id}/trace": {
      "get": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Trace",
        "operationId": "trace_agent_runs__run_id__trace_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTrace"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/agent/runs/{run_id}/recover": {
      "post": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Recover",
        "operationId": "recover_agent_runs__run_id__recover_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResult"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/agent/policies/{document_id}": {
      "get": {
        "tags": [
          "stage3-agent"
        ],
        "summary": "Policy",
        "operationId": "policy_agent_policies__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Citation"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRead"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionRead": {
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/PolicyDecision"
          },
          "authorized_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "type": "string",
            "title": "Amount",
            "default": "0.00"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "LOW"
          },
          "rule_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rule Codes"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "order_item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Item Id"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "proposed_action": {
            "$ref": "#/components/schemas/ActionType"
          },
          "final_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "execution_status": {
            "$ref": "#/components/schemas/ExecutionStatus"
          },
          "evidence_provided": {
            "type": "boolean",
            "title": "Evidence Provided"
          },
          "return_received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Received At"
          },
          "snapshot_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "rule_codes",
          "explanation",
          "policy_version",
          "id",
          "ticket_id",
          "user_id",
          "order_id",
          "order_item_id",
          "quantity",
          "idempotency_key",
          "proposed_action",
          "final_action",
          "execution_status",
          "evidence_provided",
          "return_received_at",
          "snapshot_json",
          "created_at",
          "updated_at"
        ],
        "title": "ActionRead"
      },
      "ActionType": {
        "type": "string",
        "enum": [
          "CANCEL_AND_REFUND",
          "RETURN_AND_REFUND",
          "REFUND_MISSING_ITEM",
          "REPLACE_ITEM",
          "CREATE_LOGISTICS_TICKET"
        ],
        "title": "ActionType"
      },
      "AgentStatus": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "deepseek"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps"
          },
          "max_tools": {
            "type": "integer",
            "title": "Max Tools"
          },
          "timeout_seconds": {
            "type": "number",
            "title": "Timeout Seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "configured",
          "model",
          "max_steps",
          "max_tools",
          "timeout_seconds"
        ],
        "title": "AgentStatus"
      },
      "AgentTrace": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/ChatResult"
          },
          "model_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Model Calls"
          },
          "tool_calls": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tool Calls"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run",
          "model_calls",
          "tool_calls"
        ],
        "title": "AgentTrace"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Idempotency Key"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "evidence_provided": {
            "type": "boolean",
            "title": "Evidence Provided",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "idempotency_key"
        ],
        "title": "ChatRequest"
      },
      "ChatResult": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "outcome": {
            "type": "string",
            "title": "Outcome"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          },
          "action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "llm_calls": {
            "type": "integer",
            "title": "Llm Calls"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "usage_complete": {
            "type": "boolean",
            "title": "Usage Complete"
          },
          "tool_calls": {
            "type": "integer",
            "title": "Tool Calls"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "status",
          "outcome",
          "provider",
          "model",
          "reply",
          "action",
          "citations",
          "llm_calls",
          "input_tokens",
          "output_tokens",
          "usage_complete",
          "tool_calls",
          "latency_ms",
          "error_type",
          "parent_run_id",
          "created_at",
          "finished_at"
        ],
        "title": "ChatResult"
      },
      "Citation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "version",
          "content",
          "content_hash",
          "source"
        ],
        "title": "Citation"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorRead": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorRead"
      },
...
      "ExecutionStatus": {
        "type": "string",
        "enum": [
          "BLOCKED",
          "WAITING_APPROVAL",
          "WAITING_RETURN",
          "READY",
          "SUCCESS",
          "FAILED",
          "REJECTED"
        ],
        "title": "ExecutionStatus"
      },
...
      "PolicyDecision": {
        "type": "string",
        "enum": [
          "ALLOW",
          "DENY",
          "REQUIRE_APPROVAL",
          "NEED_MORE_INFO"
        ],
        "title": "PolicyDecision"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "RiskLevel"
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "RUNNING",
          "SUCCESS",
          "FAILED"
        ],
        "title": "RunStatus"
      },
...
    }
  }
}